requirements.txt
appdirs==1.4.4
astroid==2.4.2
atomicwrites==1.4.0
attrs==20.1.0
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.3
configparser==5.0.0
crayons==0.4.0
distlib==0.3.1
filelock==3.0.12
httplib2==0.18.1
idna==2.10
iniconfig==1.0.1
isort==5.4.2
JobFunnel==2.1.9
joblib==0.16.0
lazy-object-proxy==1.4.3
lxml==4.5.2
mccabe==0.6.1
more-itertools==8.5.0
nltk==3.5
numpy==1.19.1
oauth2==1.9.0.post1
oauthlib==3.1.0
packaging==20.4
pbr==5.5.0
Pillow==8.2.0
pluggy==0.13.1
py==1.9.0
pylint==2.6.0
pyparsing==2.4.7
PyPDF2==1.26.0
pytest==6.0.1
pytest-mock==3.3.1
python-dateutil==2.8.1
PyYAML==5.3.1
regex==2020.7.14
reportlab==3.5.66
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scipy==1.5.2
selenium==3.141.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
stevedore==3.2.2
threadpoolctl==2.1.0
toml==0.10.1
tqdm==4.48.2
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
virtualenvwrapper-win==1.2.6
webdriver-manager==3.2.2
wrapt==1.12.1
